Move ClassGenerator to test folder
