+++ src/test/scala/org/kidsfirstdrc/dwh/testutils/ClassGenerator.scala
@@ -1,8 +1,8 @@
-package org.kidsfirstdrc.dwh.utils
+package org.kidsfirstdrc.dwh.testutils
 
-import org.apache.spark.sql.functions.{col, explode}
+import org.apache.spark.sql._
+import org.apache.spark.sql.functions._
 import org.apache.spark.sql.types._
-import org.apache.spark.sql.{DataFrame, Row}
 
 import java.io._
 import java.time.LocalDateTime
@@ -79,14 +79,14 @@ object ClassGenerator {
     }
 
     s"""/**
-      | * Generated by [[org.kidsfirstdrc.dwh.utils.ClassGenerator]]
-      | * on ${LocalDateTime.now()}
-      | */
-      |package $packageName
-      |
-      |$mainClass
-      |${nestedClasses.mkString("\n")}
-      |""".stripMargin
+       | * Generated by [[org.kidsfirstdrc.dwh.testutils.ClassGenerator]]
+       | * on ${LocalDateTime.now()}
+       | */
+       |package $packageName
+       |
+       |$mainClass
+       |${nestedClasses.mkString("\n")}
+       |""".stripMargin
   }
 
   def writeCLassFile(packageName: String,
@@ -113,3 +113,4 @@ object ClassGenerator {
     println(getCaseClassFileContent(packageName, className, df))
   }
 }
+
