+++ src/test/scala/org/kidsfirstdrc/dwh/testutils/external/ClinvarInput.scala
@@ -1,6 +1,6 @@
 /**
- * Generated by [[org.kidsfirstdrc.dwh.utils.ClassGenerator]]
- * on 2021-01-28T10:54:32.652
+ * Generated by [[org.kidsfirstdrc.dwh.testutils.ClassGenerator]]
+ * on 2021-01-28T14:10:00.081
  */
 package org.kidsfirstdrc.dwh.testutils.external
