+++ src/test/scala/org/kidsfirstdrc/dwh/testutils/external/ClinvarOutput.scala
@@ -1,6 +1,6 @@
 /**
- * Generated by [[org.kidsfirstdrc.dwh.utils.ClassGenerator]]
- * on 2021-01-28T10:54:37.742
+ * Generated by [[org.kidsfirstdrc.dwh.testutils.ClassGenerator]]
+ * on 2021-01-28T14:10:03.272
  */
 package org.kidsfirstdrc.dwh.testutils.external
